+++ Machampion/MainMenuState.cpp
@@ -29,7 +29,7 @@ void MainMenuState::update() {
 	//If state is done loading and game objects are not empty
 	if (m_loadingComplete && !m_gameObjects.empty()) {
 		//If there is a joystick and joysticks have not already been initialized
-		if (SDL_NumJoysticks > 0 && !TheInputHandler::Instance()->joysticksInitialized()) {
+		if ((int)SDL_NumJoysticks > 0 && !TheInputHandler::Instance()->joysticksInitialized()) {
 			//Initialize joysticks
 			TheInputHandler::Instance()->initializeJoysticks();
 		}
